Removing warnings.
Also fixes two cases where a member variable was being used instead of a parameter.

+++ biojava-structure/src/main/java/org/biojava/nbio/structure/contact/StructureInterfaceList.java
@@ -277,7 +277,7 @@ public class StructureInterfaceList implements Serializable, Iterable<StructureI
 				}
 			}
 			if (countPairs>0) { 
-				averageScore = averageScore/(double)countPairs;
+				averageScore = averageScore/countPairs;
 			} else {
 				// if only one interface in cluster we set the score to the maximum
 				averageScore = 1.0;
